+++ src/api/java/blusunrize/immersiveengineering/api/Lib.java
@@ -77,6 +77,7 @@ public class Lib
 	public static final String GUIID_RedstoneProbe = "redstoneprobe";
 	public static final String GUIID_RedstoneStateCell = "redstonestatecell";
 	public static final String GUIID_RedstoneTimer = "redstonetimer";
+	public static final String GUIID_RedstoneSwitchboard = "redstoneswitchboard";
 	//Items
 	public static final String GUIID_Revolver = "revolver";
 	public static final String GUIID_Toolbox = "toolbox";