+++ changelog.md
@@ -1,6 +1,7 @@
 ##### Version 1.20.4-11.5.1-pre
 - Add the redstone state cell for storing signals (BluSunrize)
 - Add the redstone timer for emitting regular signals (BluSunrize)
+- Add the redstone switchboard for remapping redstone signals between networks (BluSunrize)
 - Fix waterlogged posts not dropping items (BluSunrize)
 - Fix treated windows and feedthrough insulators letting rain through (BluSunrize, voidsong-dragonfly)
 - Fix crates not syncing their names to the client (BluSunrize)